// src/app/utils/data/models/event.model.ts
export interface Event {
    categories: string[];
    dateEndDay: number;
    dateEndMillis: number;
    dateEndMonth: number;
    dateEndYear: number;
    dateStartDay: number;
    dateStartMillis: number;
    dateStartMonth: number;
    dateStartYear: number;
    filters: string[];
    location: string;
    name: string;
    price: number;
    shortLocation: string;
}